moved PLY load options from oglPlyHelper to new contents/pointcloud.ts. This removed the last dependency of RenderingEngine on OGL-specific stuff.
(cherry picked from commit 42a9648eff8b8b8cf6cbe513343b19730718b94b)
